requirements.txt
opencv-python
numpy
tensorflow-cpu
pandas
fastapi
PyYAML
uvicorn
python-multipart




